Changed the type of the memory_size field for GPUs, to account for non integer values

+++ boaviztapi/dto/component_dto.py
@@ -90,7 +90,7 @@ class Gpu(BaseModel):
     model: Optional[str] = None
     architecture: Optional[str] = None
     tdp: Optional[int] = None
-    memory_size: Optional[int] = None
+    memory_size: Optional[float] = None
     memory_density: Optional[float] = None
 
     def to_component(self):
